+++ mne/time_frequency/__init__.py
@@ -1,6 +1,6 @@
 """Time frequency analysis tools
 """
 
-from .tfr import induced_power, single_trial_power
+from .tfr import induced_power, single_trial_power, morlet
 from .psd import compute_raw_psd
-from .ar import yule_walker, ar_raw
+from .ar import yule_walker, ar_raw, fir_filter_raw